ADD: Added Product's validations.

+++ OnlineStore/Services/ProductService.cs
@@ -2,7 +2,7 @@
 using OnlineStore.Domain.Models;
 using OnlineStore.Domain.Respsitories;
 using OnlineStore.Domain.Services;
-using OnlineStore.Dtos;
+using OnlineStore.Domain.Dtos;
 using System.Collections.Generic;
 using System.Linq;
 using System.Transactions;
@@ -14,31 +14,56 @@ namespace OnlineStore.Services
 
 		public ProductService(
 			IProductRepository productRepository,
+			IProductGroupRepository productGroupRepository,
 			IProductsDeliveryGroupsRepository productsDeliveryGroupsRepository,
 			ITariffRepository tariffRepository,
 			IUnitOfWork unitOfWork)
 		{
 			_productRepository = productRepository;
+			_productGroupRepository = productGroupRepository;
 			_productsDeliveryGroupsRepository = productsDeliveryGroupsRepository;
 			_tariffRepository = tariffRepository;
 			_unitOfWork = unitOfWork;
 		}
 		private readonly IProductRepository _productRepository;
+		private readonly IProductGroupRepository _productGroupRepository;
 		private readonly IProductsDeliveryGroupsRepository _productsDeliveryGroupsRepository;
 		private readonly ITariffRepository _tariffRepository;
 		private readonly IUnitOfWork _unitOfWork;
 
 		public void Delete(Product product)
 		{
-			_productRepository.Delete(product);
-			_unitOfWork.Commit();
+			if (DeleteValidate(product, out List<string> messagaes))
+			{
+				using (var scope = new TransactionScope(TransactionScopeOption.Required, new TransactionOptions { IsolationLevel = IsolationLevel.ReadCommitted }))
+				{
+					_productRepository.Delete(product);
+					var productsDeliveryGroups = _productsDeliveryGroupsRepository.GetByPtoduct(product.Id)?.ToList();
+					_productsDeliveryGroupsRepository.DeleteRange(productsDeliveryGroups);
+					_unitOfWork.Commit();
+					scope.Complete();
+				}
+			}
+			else
+				throw new System.Exception(string.Join(",", messagaes));
 		}
 
 		public void Delete(int id)
 		{
-
-			_productRepository.Delete(id);
-			_unitOfWork.Commit();
+			var product = this.Get(id);
+			if (DeleteValidate(product, out List<string> messagaes))
+			{
+				using (var scope = new TransactionScope(TransactionScopeOption.Required, new TransactionOptions { IsolationLevel = IsolationLevel.ReadCommitted }))
+				{
+					_productRepository.Delete(id);
+					var productsDeliveryGroups = _productsDeliveryGroupsRepository.GetByPtoduct(product.Id)?.ToList();
+					_productsDeliveryGroupsRepository.DeleteRange(productsDeliveryGroups);
+					_unitOfWork.Commit();
+					scope.Complete();
+				}
+			}
+			else
+				throw new System.Exception(string.Join(",", messagaes));
 		}
 
 		public Product Get(int id)
@@ -60,28 +85,35 @@ namespace OnlineStore.Services
 
 		public void Submit(Product product)
 		{
-			using (var scope = new TransactionScope(TransactionScopeOption.Required, new TransactionOptions { IsolationLevel = IsolationLevel.ReadCommitted }))
+			if (SubmitValidate(product, out List<string> messagaes))
 			{
-				if (product.Id > 0)
+				using (var scope = new TransactionScope(TransactionScopeOption.Required, new TransactionOptions { IsolationLevel = IsolationLevel.ReadCommitted }))
 				{
-					var oldData = _productsDeliveryGroupsRepository.GetByPtoduct(product.Id)?.ToList();
-					_productsDeliveryGroupsRepository.DeleteRange(oldData);
-					_unitOfWork.Commit();
-					_productsDeliveryGroupsRepository.InsertRange(product.ProductsDeliveryGroups);
-					_productRepository.Update(product);
-				}
-				else
-				{
-					_productRepository.Insert(product);
-					_unitOfWork.Commit();
+					if (product.Id > 0)
+					{
 
-					product.ProductsDeliveryGroups.Select(item => item.ProductId = product.Id);
-					_productsDeliveryGroupsRepository.InsertRange(product.ProductsDeliveryGroups);
-					
+						var oldData = _productsDeliveryGroupsRepository.GetByPtoduct(product.Id)?.ToList();
+						_productsDeliveryGroupsRepository.DeleteRange(oldData);
+						_unitOfWork.Commit();
+						_productsDeliveryGroupsRepository.InsertRange(product.ProductsDeliveryGroups);
+						_productRepository.Update(product);
+
+					}
+					else
+					{
+						_productRepository.Insert(product);
+						_unitOfWork.Commit();
+
+						product.ProductsDeliveryGroups.Select(item => item.ProductId = product.Id);
+						_productsDeliveryGroupsRepository.InsertRange(product.ProductsDeliveryGroups);
+
+					}
+					_unitOfWork.Commit();
+					scope.Complete();
 				}
-				_unitOfWork.Commit();
-				scope.Complete();
 			}
+			else
+				throw new System.Exception(string.Join(",", messagaes));
 		}
 
 		public int GetMinimumProductTariff(int productId)
@@ -92,6 +124,33 @@ namespace OnlineStore.Services
 		{
 			return _tariffRepository.GetProductTariffList(ids);
 		}
-		
+
+		private bool SubmitValidate(Product product, out List<string> messagaes)
+		{
+			messagaes = new List<string>();
+
+			if (string.IsNullOrEmpty(product.Name))
+				messagaes.Add("Product Name Can't be emty.");
+
+			if (product.ProductGroupId < 1)
+				messagaes.Add("Product Group Can't be emty.");
+
+			if (product.RegisterDate == null)
+				messagaes.Add("Product Register Date Can't be emty.");
+
+			if (!_productGroupRepository.ProductGroupExist(product.ProductGroupId))
+				messagaes.Add("Product Group Is Invalid.");
+
+			if (_productRepository.IsNameDuplicated(product))
+				messagaes.Add("Product Name Is Duplicate.");
+
+			return messagaes.Count < 1;
+		}
+
+		private bool DeleteValidate(Product product, out List<string> messagaes)
+		{
+			messagaes = new List<string>();
+			return true;
+		}
 	}
 }
